Bump all the things and simplify data updates

+++ lib/tasks/sptrans_import.rake
@@ -1,8 +1,13 @@
 namespace :sptrans do
-  desc 'Imports data downloaded from http://www.sptrans.com.br/desenvolvedores/ (renamed gtfs-sptrans.zip)'
+  desc 'Imports data downloaded from SPTrans'
   task :import do
+    FileUtils.mkdir_p "tmp/cache"
+    next unless system("echo 'Downloading...'; wget https://openmobilitydata.org/p/sptrans/1049/latest/download -O/tmp/gtfs-sptrans.zip")
+
     Rake::Task["db:reset"].invoke
     Rake::Task["gtfs_engine:update"].invoke
     Rake::Task["geohashes:rehash"].invoke
+
+    File.delete("/tmp/gtfs-sptrans.zip")
   end
 end
